refactor: remove unused components for topology, moved to integrations
This all logic was moved to integrations hence no need for it

fix: fix failing builds

fix: move to integrations directory
